+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Component Enhancements
 
 * `DateRange`: Two new props have been added, `startDateProps` and `endDateProps`, to apply props to the child `Date` components.
+* `MultiStepWizard` now allows adding callbacks when clicking on Next/Back button and allows adding validation callback before wizard submission when clicking on Submit button. 
 
 # 0.33.2